+++ .gitignore
@@ -9,10 +9,15 @@ DerivedData/
 .netrc
 *.class
 bin/
+BuildLogic/out/
 
 # Ignore gradle build artifacts
 .gradle
 **/build/
+lib/
+
+# Ignore package resolved
+Package.resolved
 
 # Avoid ignoring Gradle wrapper jar file (.jar files are usually ignored)
 !gradle-wrapper.jar
@@ -24,4 +29,4 @@ bin/
 .gradletasknamecache
 
 # Ignore generated sources
-JavaSwiftKitDemo/src/main/java/com/example/swift/generated/*
+**/generated/
