+++ BuildLogic/src/main/kotlin/build-logic.java-common-conventions.gradle.kts
@@ -12,6 +12,8 @@
 //
 //===----------------------------------------------------------------------===//
 
+import java.util.*
+
 plugins {
     java
 }
@@ -40,16 +42,39 @@ tasks.withType(JavaCompile::class).forEach {
     it.options.compilerArgs.add("-Xlint:preview")
 }
 
+
+fun javaLibraryPaths(): List<String> {
+    val osName = System.getProperty("os.name")
+    val osArch = System.getProperty("os.arch")
+    val isLinux = osName.lowercase(Locale.getDefault()).contains("linux")
+
+    return listOf(
+        if (osName.lowercase(Locale.getDefault()).contains("linux")) {
+            """$rootDir/.build/$osArch-unknown-linux-gnu/debug/"""
+        } else {
+            if (osArch.equals("aarch64")) {
+                """$rootDir/.build/arm64-apple-macosx/debug/"""
+            } else {
+                """$rootDir/.build/$osArch-apple-macosx/debug/"""
+            }
+        },
+        if (isLinux) {
+            "/usr/lib/swift/linux"
+        } else {
+            // assume macOS
+            "/usr/lib/swift/"
+        }
+    )
+}
+
+
 // Configure paths for native (Swift) libraries
 tasks.test {
     jvmArgs(
         "--enable-native-access=ALL-UNNAMED",
 
         // Include the library paths where our dylibs are that we want to load and call
-        "-Djava.library.path=" + listOf(
-            """$rootDir/.build/arm64-apple-macosx/debug/""",
-            "/usr/lib/swift/"
-        ).joinToString(File.pathSeparator)
+        "-Djava.library.path=" + javaLibraryPaths().joinToString(File.pathSeparator)
     )
 }
 
@@ -60,14 +85,15 @@ tasks.withType<Test> {
 }
 
 
-val buildSwiftJExtract = tasks.register<Exec>("buildSwiftJExtract") {
-    description = "Builds Swift targets, including jextract-swift"
-
-    workingDir("..")
-    commandLine("make")
-}
-
-tasks.build {
-    dependsOn(buildSwiftJExtract)
-}
+// TODO: This is a crude workaround, we'll remove 'make' soon and properly track build dependencies
+// val buildSwiftJExtract = tasks.register<Exec>("buildMake") {
+//    description = "Triggers 'make' build"
+//
+//    workingDir(rootDir)
+//    commandLine("make")
+// }
+//
+// tasks.build {
+//     dependsOn(buildSwiftJExtract)
+// }
