+++ Makefile
@@ -37,17 +37,15 @@ BUILD_DIR := .build/$(ARCH_SUBDIR)-apple-macosx
 LIB_SUFFIX := dylib
 endif
 
+SAMPLES_DIR := "Samples"
 
 all: generate-all
 
-$(BUILD_DIR)/debug/libJavaKit.$(LIB_SUFFIX) $(BUILD_DIR)/debug/libJavaKitExample.$(LIB_SUFFIX) $(BUILD_DIR)/debug/Java2Swift:
+$(BUILD_DIR)/debug/libJavaKit.$(LIB_SUFFIX) $(BUILD_DIR)/debug/Java2Swift:
 	swift build
 
-./JavaSwiftKitDemo/build/classes/java/main/com/example/swift/HelloSubclass.class: JavaSwiftKitDemo/src/main/java/com/example/swift
-	./gradlew build
-
-run: $(BUILD_DIR)/debug/libJavaKit.$(LIB_SUFFIX) $(BUILD_DIR)/debug/libJavaKitExample.$(LIB_SUFFIX) JavaSwiftKitDemo/src/main/java/com/example/swift
-	java -cp JavaSwiftKitDemo/build/classes/java/main -Djava.library.path=$(BUILD_DIR)/debug/ com.example.swift.HelloSwift
+run: $(BUILD_DIR)/debug/libJavaKit.$(LIB_SUFFIX) $(BUILD_DIR)/debug/libExampleSwiftLibrary.$(LIB_SUFFIX)
+	./gradlew Samples:JavaKitSampleApp:run
 
 Java2Swift: $(BUILD_DIR)/debug/Java2Swift
 
@@ -70,7 +68,8 @@ generate-JavaKitNetwork: Java2Swift generate-JavaKit
 generate-all: generate-JavaKit generate-JavaKitReflection generate-JavaKitJar generate-JavaKitNetwork \
 			  jextract-swift
 clean:
-	rm -rf .build
+	rm -rf .build; \
+	rm -rf Samples/SwiftKitExampleApp/src/generated/java/*
 
 format:
 	swift format --recursive . -i
@@ -81,7 +80,6 @@ format:
 
 JEXTRACT_BUILD_DIR="$(BUILD_DIR)/jextract"
 
-# Parameter: Swift source file
 define make_swiftinterface
     $(eval $@_MODULE = $(1))
     $(eval $@_FILENAME = $(2))
@@ -99,17 +97,21 @@ jextract-swift: generate-JExtract-interface-files
 
 generate-JExtract-interface-files: $(BUILD_DIR)/debug/libJavaKit.$(LIB_SUFFIX)
 	echo "Generate .swiftinterface files..."
-	@$(call make_swiftinterface, "JavaKitExample", "MySwiftLibrary")
-	@$(call make_swiftinterface, "JavaKitExample", "SwiftKit")
+	@$(call make_swiftinterface, "ExampleSwiftLibrary", "MySwiftLibrary")
+	@$(call make_swiftinterface, "SwiftKitSwift", "SwiftKit")
 
 jextract-run: jextract-swift generate-JExtract-interface-files
 	swift run jextract-swift  \
 		--package-name com.example.swift.generated \
-		--swift-module JavaKitExample \
-		--output-directory JavaSwiftKitDemo/src/main/java \
-		$(BUILD_DIR)/jextract/JavaKitExample/MySwiftLibrary.swiftinterface \
-		$(BUILD_DIR)/jextract/JavaKitExample/SwiftKit.swiftinterface
+		--swift-module ExampleSwiftLibrary \
+		--output-directory ${SAMPLES_DIR}/SwiftKitSampleApp/src/generated/java \
+		$(BUILD_DIR)/jextract/ExampleSwiftLibrary/MySwiftLibrary.swiftinterface; \
+	swift run jextract-swift \
+		--package-name org.swift.swiftkit.generated \
+		--swift-module SwiftKitSwift \
+		--output-directory ${SAMPLES_DIR}/SwiftKitSampleApp/src/generated/java \
+		$(BUILD_DIR)/jextract/SwiftKitSwift/SwiftKit.swiftinterface
 
 
 jextract-run-java: jextract-swift generate-JExtract-interface-files
-	./gradlew run
+	./gradlew Samples:SwiftKitSampleApp:run
