--- SwiftJavaKitExample/README.md
@@ -1,2 +1,0 @@
-# Gradle build for Sources/JavaKitExample
-
