+++ SwiftKit/src/main/java/org/swift/swiftkit/SwiftHeapObject.java
