+++ buildSrc/build.gradle
@@ -1,0 +1,3 @@
+repositories {
+    mavenCentral()
+}
