+++ buildSrc/src/main/groovy/org/swift/swiftkit/gradle/BuildUtils.groovy
@@ -1,0 +1,36 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the Swift.org open source project
+//
+// Copyright (c) 2024 Apple Inc. and the Swift.org project authors
+// Licensed under Apache License v2.0
+//
+// See LICENSE.txt for license information
+// See CONTRIBUTORS.txt for the list of Swift.org project authors
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+//===----------------------------------------------------------------------===//
+
+package org.swift.swiftkit.gradle
+
+final class BuildUtils {
+    static def javaLibraryPaths(File rootDir) {
+        def osName = System.getProperty("os.name")
+        def osArch = System.getProperty("os.arch")
+        def isLinux = osName.toLowerCase(Locale.getDefault()).contains("linux")
+
+        return [
+                osName.toLowerCase(Locale.getDefault()).contains("linux") ?
+                        "${rootDir}/.build/${osArch}-unknown-linux-gnu/debug/" :
+                        osArch == "aarch64" ?
+                                "${rootDir}/.build/arm64-apple-macosx/debug/" :
+                                "${rootDir}/.build/${osArch}-apple-macosx/debug/",
+                isLinux ?
+                        "/usr/lib/swift/linux" :
+                        // assume macOS
+                        "/usr/lib/swift/"
+        ]
+    }
+
+}
