+++ settings.gradle
@@ -1,0 +1,29 @@
+//===----------------------------------------------------------------------===//
+//
+// This source file is part of the Swift.org open source project
+//
+// Copyright (c) 2024 Apple Inc. and the Swift.org project authors
+// Licensed under Apache License v2.0
+//
+// See LICENSE.txt for license information
+// See CONTRIBUTORS.txt for the list of Swift.org project authors
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+//===----------------------------------------------------------------------===//
+
+pluginManagement{
+    includeBuild("BuildLogic")
+}
+
+rootProject.name = "swift-java"
+
+include "SwiftKit"
+
+// Include sample apps -- you can run them via `gradle Name:run`
+new File(rootDir, "Samples").listFiles().each {
+    if (it.directory && new File(it, 'build.gradle').exists()) {
+        include ":Samples:${it.name}"
+    }
+}
+
